user auth

+++ src/components/WorkoutDetails.js
@@ -1,19 +1,28 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
-        const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+        if (!user) {
+            return
+        }
+
+        const response = await fetch('https://excercise-buddy-app-backend.vercel.app/api/workouts/' + workout._id, {
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
 
-        if (response.ok){
-            dispatch({type: 'DELETE_WORKOUT', payload: json})
+        if (response.ok) {
+            dispatch({ type: 'DELETE_WORKOUT', payload: json })
         }
     }
 
@@ -23,7 +32,7 @@ const WorkoutDetails = ({ workout }) => {
             <p><strong>Load (kg): </strong>{workout.load}</p>
             <p><strong>Number of reps: </strong>{workout.reps}</p>
             <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
-      <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
+            <span className="material-symbols-outlined" onClick={handleClick}>delete</span>
         </div>
     )
 }
